+++ backend/src/main/java/wooteco/prolog/login/ui/auto_interceptor_register/scanner/MappingAnnotation.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.login.ui.auto_interceptor_register;
+package wooteco.prolog.login.ui.auto_interceptor_register.scanner;
 
 import org.springframework.web.bind.annotation.*;
