+++ backend/src/main/java/wooteco/prolog/login/ui/auto_interceptor_register/scanner/URIScanner.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.login.ui.auto_interceptor_register;
+package wooteco.prolog.login.ui.auto_interceptor_register.scanner;
 
 import org.springframework.util.StringUtils;
