+++ backend/src/test/java/wooteco/prolog/login/ui/auto_interceptor_register/MappingAnnotationTest.java
@@ -7,16 +7,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import wooteco.prolog.login.domain.AuthMemberPrincipal;
+import wooteco.prolog.login.ui.auto_interceptor_register.scanner.MappingAnnotation;
+import wooteco.prolog.login.ui.auto_interceptor_register.scanner.MethodScanner;
 
 import java.lang.reflect.Member;
 import java.lang.reflect.Method;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class MappingAnnotationTest {
