+++ backend/src/test/java/wooteco/prolog/login/ui/auto_interceptor_register/MethodScannerTest.java
@@ -3,6 +3,7 @@ package wooteco.prolog.login.ui.auto_interceptor_register;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.prolog.login.domain.AuthMemberPrincipal;
+import wooteco.prolog.login.ui.auto_interceptor_register.scanner.MethodScanner;
 import wooteco.prolog.login.ui.auto_interceptor_register.test_classes.ControllerClass;
 import wooteco.prolog.login.ui.auto_interceptor_register.test_classes.RestControllerClass;
